+++ modules/basegl/include/modules/basegl/processors/imageprocessing/imagemapping.h
@@ -38,23 +38,7 @@
 namespace inviwo {
 class TextureUnitContainer;
 
-/** \docpage{org.inviwo.ImageMapping, Image Mapping}
- * Maps the input image to an output image with the help of a transfer function.
- * ![](org.inviwo.ImageMapping.png?classIdentifier=org.inviwo.ImageMapping)
- *
- * ### Inports
- *   * __ImageInport__ The input image.
- *
- * ### Outports
- *   * __ImageOutport__ The output image.
- *
- * ### Properties
- *   * __Transfer Function__ The transfer function used for mapping input to output values
- *                           including the alpha channel.
- */
-
-/*! \class ImageMapping
- *
+/**
  * \brief Maps the input image to an output with the help of a transfer function.
  */
 class IVW_MODULE_BASEGL_API ImageMapping : public ImageGLProcessor {