+++ modules/basegl/src/processors/imageprocessing/imageglprocessor.cpp
@@ -55,13 +55,49 @@ namespace inviwo {
 class DataFormatBase;
 class ShaderResource;
 
+namespace detail {
+
+void syncMetaDataAndColorLayerState(ImageInport& inport, Image* dstImage,
+                                    std::optional<SwizzleMask> swizzleMask = std::nullopt) {
+    const auto srcImage = inport.getData();
+    const auto srcColorLayer = srcImage->getColorLayer();
+    auto dstColorLayer = dstImage->getColorLayer();
+
+    dstImage->copyMetaDataFrom(*srcImage);
+
+    dstColorLayer->setSwizzleMask(swizzleMask.value_or(srcColorLayer->getSwizzleMask()));
+    dstColorLayer->setWrapping(inport.getData()->getColorLayer()->getWrapping());
+    dstColorLayer->setInterpolation(inport.getData()->getColorLayer()->getInterpolation());
+}
+
+bool equalColorLayerState(ImageInport& inport, ImageOutport& outport,
+                          std::optional<SwizzleMask> swizzleMask = std::nullopt) {
+    const auto source = inport.getData()->getColorLayer();
+    auto dest = outport.getData()->getColorLayer();
+
+    return swizzleMask.value_or(source->getSwizzleMask()) == dest->getSwizzleMask() &&
+           source->getWrapping() == dest->getWrapping() &&
+           source->getInterpolation() == dest->getInterpolation();
+}
+
+bool needsNewImage(ImageInport& inport, ImageOutport& outport, const size2_t& dstDim,
+                   const DataFormatBase* dstDataFormat,
+                   std::optional<SwizzleMask> swizzleMask = std::nullopt) {
+
+    return dstDataFormat != outport.getData()->getDataFormat() ||
+           dstDim != outport.getData()->getDimensions() ||
+           !equalColorLayerState(inport, outport, swizzleMask);
+}
+
+}  // namespace detail
+
 ImageGLProcessor::ImageGLProcessor(const std::string& fragmentShader, bool buildShader)
     : ImageGLProcessor(utilgl::findShaderResource(fragmentShader), buildShader) {}
 
 ImageGLProcessor::ImageGLProcessor(std::shared_ptr<const ShaderResource> fragmentShader,
                                    bool buildShader)
     : Processor()
-    , inport_("inputImage", true)
+    , inport_("inputImage", "The input image"_help, OutportDeterminesSize::Yes)
     , outport_("outputImage", false)
     , dataFormat_(nullptr)
     , swizzleMask_(swizzlemasks::rgba)
@@ -71,6 +107,7 @@ ImageGLProcessor::ImageGLProcessor(std::shared_ptr<const ShaderResource> fragmen
 
     addPort(inport_);
     addPort(outport_);
+    outport_.setHelp("The output image"_help);
 
     inport_.onChange([this]() {
         markInvalid();
@@ -92,10 +129,15 @@ void ImageGLProcessor::process() {
         internalInvalid_ = false;
 
         const size2_t dim(calcOutputDimensions());
-        if (dataFormat_) {
-            createCustomImage(dim, dataFormat_, swizzleMask_, inport_, outport_);
-        } else {
-            createDefaultImage(dim, inport_, outport_);
+        auto dataformat = getDestinationDataFormat();
+        auto swizzlemask = getDestinationSwizzleMask();
+
+        if (!outport_.hasEditableData() ||
+            detail::needsNewImage(inport_, outport_, dim, dataformat, swizzlemask)) {
+            Image* img = new Image(dim, dataformat);
+            detail::syncMetaDataAndColorLayerState(inport_, img, swizzlemask);
+
+            outport_.setData(img);
         }
     }
 
@@ -129,40 +171,6 @@ void ImageGLProcessor::postProcess() {}
 
 void ImageGLProcessor::afterInportChanged() {}
 
-void ImageGLProcessor::createCustomImage(const size2_t& dim, const DataFormatBase* dataFormat,
-                                         const SwizzleMask& swizzleMask, ImageInport& inport,
-                                         ImageOutport& outport) {
-
-    if (!outport.hasEditableData() || dataFormat != outport.getData()->getDataFormat() ||
-        dim != outport.getData()->getDimensions()) {
-        Image* img = new Image(dim, dataFormat);
-        img->copyMetaDataFrom(*inport.getData());
-        img->getColorLayer()->setSwizzleMask(swizzleMask);
-        outport.setData(img);
-    } else if (outport.hasEditableData() &&
-               outport.getData()->getColorLayer()->getSwizzleMask() != swizzleMask) {
-        outport.getEditableData()->getColorLayer()->setSwizzleMask(swizzleMask);
-    }
-}
-
-void ImageGLProcessor::createDefaultImage(const size2_t& dim, ImageInport& inport,
-                                          ImageOutport& outport) {
-    const DataFormatBase* format = inport.getData()->getDataFormat();
-
-    const auto swizzleMask = inport.getData()->getColorLayer()->getSwizzleMask();
-
-    if (!outport.hasEditableData() || format != outport.getData()->getDataFormat() ||
-        dim != outport.getData()->getDimensions() ||
-        swizzleMask != outport.getData()->getColorLayer()->getSwizzleMask()) {
-        Image* img = new Image(dim, format);
-        img->copyMetaDataFrom(*inport.getData());
-        // forward swizzle mask of the input
-        img->getColorLayer()->setSwizzleMask(swizzleMask);
-
-        outport.setData(img);
-    }
-}
-
 size2_t ImageGLProcessor::calcOutputDimensions() const {
     size2_t dimensions;
     if (outport_.isHandlingResizeEvents() || !inport_.isOutportDeterminingSize()) {
@@ -171,6 +179,24 @@ size2_t ImageGLProcessor::calcOutputDimensions() const {
         dimensions = inport_.getData()->getDimensions();
     }
     return dimensions;
+}
+
+const DataFormatBase* ImageGLProcessor::getDestinationDataFormat() const {
+    if (dataFormat_) {
+        return dataFormat_;
+    } else if (inport_.hasData()) {
+        return inport_.getData()->getDataFormat();
+    }
+    return DataVec4UInt8::get();
+}
+
+SwizzleMask ImageGLProcessor::getDestinationSwizzleMask() const {
+    if (dataFormat_) {
+        return swizzleMask_;
+    } else if (inport_.hasData()) {
+        return inport_.getData()->getColorLayer()->getSwizzleMask();
+    }
+    return swizzlemasks::rgba;
 }
 
 }  // namespace inviwo