+++ modules/python3/src/python3module.cpp
@@ -36,21 +36,21 @@
 #include <inviwo/core/datastructures/datarepresentation.h>           // for DataRepresentation<>...
 #include <inviwo/core/datastructures/representationconverter.h>      // for RepresentationConver...
 #include <inviwo/core/datastructures/representationfactoryobject.h>  // for RepresentationFactor...
+#include <inviwo/core/datastructures/volume/volume.h>
 #include <inviwo/core/datastructures/volume/volumerepresentation.h>  // for VolumeRepresentation
-#include <inviwo/core/util/commandlineparser.h>                      // for CommandLineParser
-#include <inviwo/core/util/exception.h>                              // for ModuleInitException
-#include <inviwo/core/util/filesystem.h>                             // for fileExists
-#include <inviwo/core/util/logcentral.h>                             // for LogCentral, LogWarn
-#include <inviwo/core/util/pathtype.h>                               // for PathType, PathType::...
-#include <inviwo/core/util/sourcecontext.h>                          // for IVW_CONTEXT
-#include <modules/python3/processors/numpymandelbrot.h>              // for NumpyMandelbrot
-#include <modules/python3/processors/numpymeshcreatetest.h>          // for NumPyMeshCreateTest
-#include <modules/python3/processors/numpyvolume.h>                  // for NumPyVolume
-#include <modules/python3/processors/pythonscriptprocessor.h>        // for PythonScriptProcessor
-#include <modules/python3/pythoninterpreter.h>                       // for PythonInterpreter
-#include <modules/python3/pythonlogger.h>                            // for PythonLogger
-#include <modules/python3/pythonscript.h>                            // for PythonScriptDisk
-#include <modules/python3/volumepy.h>                                // for VolumePy, VolumePy2R...
+#include <inviwo/core/datastructures/image/layer.h>
+#include <inviwo/core/datastructures/image/layerrepresentation.h>
+#include <inviwo/core/util/commandlineparser.h>  // for CommandLineParser
+#include <inviwo/core/util/exception.h>          // for ModuleInitException
+#include <inviwo/core/util/filesystem.h>         // for fileExists
+#include <inviwo/core/util/logcentral.h>         // for LogCentral, LogWarn
+#include <inviwo/core/util/pathtype.h>           // for PathType, PathType::...
+#include <inviwo/core/util/sourcecontext.h>      // for IVW_CONTEXT
+#include <modules/python3/pythonscript.h>
+#include <modules/python3/pythoninterpreter.h>  // for PythonInterpreter
+#include <modules/python3/pythonlogger.h>       // for PythonLogger
+#include <modules/python3/volumepy.h>           // for VolumePy, VolumePy2R...
+#include <modules/python3/layerpy.h>
 
 #include <tclap/ArgException.h>  // for ArgParseException
 #include <tclap/ValueArg.h>      // for ValueArg
@@ -76,6 +76,19 @@ public:
             volume->getInterpolation(), volume->getWrapping());
 
         return volumePy;
+    }
+};
+
+class LayerPyFactoryObject
+    : public RepresentationFactoryObjectTemplate<LayerRepresentation, LayerPy> {
+public:
+    virtual std::unique_ptr<LayerRepresentation> create(
+        const typename LayerRepresentation::ReprOwner* layer) {
+        auto layerPy = std::make_unique<LayerPy>(layer->getDimensions(), layer->getLayerType(),
+                                                 layer->getDataFormat(), layer->getSwizzleMask(),
+                                                 layer->getInterpolation(), layer->getWrapping());
+
+        return layerPy;
     }
 };
 }  // namespace
@@ -111,10 +124,10 @@ Python3Module::Python3Module(InviwoApplication* app)
     registerRepresentationConverter<VolumeRepresentation>(
         std::make_unique<VolumePy2RAMConverter>());
 
-    registerProcessor<NumPyVolume>();
-    registerProcessor<NumpyMandelbrot>();
-    registerProcessor<NumPyMeshCreateTest>();
-    registerProcessor<PythonScriptProcessor>();
+    registerRepresentationFactoryObject<LayerRepresentation>(
+        std::make_unique<LayerPyFactoryObject>());
+    registerRepresentationConverter<LayerRepresentation>(std::make_unique<LayerRAM2PyConverter>());
+    registerRepresentationConverter<LayerRepresentation>(std::make_unique<LayerPy2RAMConverter>());
 
     // We need to import inviwopy to trigger the initialization code in inviwopy.cpp, this is needed
     // to be able to cast cpp/inviwo objects to python objects.